% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{criterionC}
\alias{criterionC}
\title{Calculate group version Cp criterion}
\usage{
criterionC(X, y, sigma, betas, g_index, grps, K)
}
\arguments{
\item{X}{X}

\item{y}{y}

\item{sigma}{standard error for y}

\item{betas}{paths of coefficients}

\item{g_index}{group id}

\item{grps}{unique groups id}

\item{K}{inter-group norm}
}
\value{
a path length criterion
}
\description{
Use Rcpp for calculating Cp criterion
}
\examples{
\dontrun{
library(SA23204165)
library(MASS)
rho = 0.2
X <- mvrnorm(200,mu = rep(0,10), Sigma = diag(10) + rho*(1-diag(10)))
beta <- c(0.2,0.4,0.5,0.6,0.6,0.6,0.2,0.3,0.5,0.7)
g.index <- c(1,1,1,2,2,2,3,3,4,5)
y <- X\%*\%beta + rnorm(200)
fit <- glars(X,y,g.index)
criterionC(fit$X, fit$y, sd(fit$y), fit$betas, g.index, unique(g.index),c(unname(table(g.index))))
}
}
