% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glars.R
\name{glars}
\alias{glars}
\title{A function for constructing path of group LARS}
\usage{
glars(
  X,
  y,
  g.index,
  K = c(unname(table(g.index))),
  normlized = TRUE,
  gram = TRUE
)
}
\arguments{
\item{X}{The observed patients number}

\item{y}{The  features dimension of covariates}

\item{g.index}{The pairwise correlation in design matrix}

\item{K}{Inter-group norm, usually take p_jI_j.}

\item{normlized}{The pairwise correlation type of features}

\item{gram}{The true coefficient provided}
}
\value{
a list with generated coefficient path and add path
}
\description{
A method for constructing path of group LARS
}
\examples{
\dontrun{
library(SA23204165)
library(MASS)
rho = 0.2
X <- mvrnorm(200,mu = rep(0,10), Sigma = diag(10) + rho*(1-diag(10)))
beta <- c(0.2,0.4,0.5,0.6,0.6,0.6,0.2,0.3,0.5,0.7)
g.index <- c(1,1,1,2,2,2,3,3,4,5)
y <- X\%*\%beta + rnorm(200)
fit <- glars(X,y,g.index)
fit$betas
}
}
