% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path.R
\name{path}
\alias{path}
\title{plot fitting path}
\usage{
path(fit, xrange = c(0, 1), yrange = c(0, 1))
}
\arguments{
\item{fit}{return of g.LARS function}

\item{xrange}{xlabel range}

\item{yrange}{ylabel range}
}
\value{
a list with generated coefficient path and add path
}
\description{
A Gibbs sampler using R
}
\examples{
\dontrun{
library(SA23204165)
library(MASS)
rho = 0.2
X <- mvrnorm(200,mu = rep(0,10), Sigma = diag(10) + rho*(1-diag(10)))
beta <- c(0.2,0.4,0.5,0.6,0.6,0.6,0.2,0.3,0.5,0.7)
g.index <- c(1,1,1,2,2,2,3,3,4,5)
y <- X\%*\%beta + rnorm(200)
fit <- glars(X,y,g.index)
path(fit)
}
}
