\name{SA23204165}
\alias{SA23204165}
\docType{package}
\title{This package provides several functions for benchmarking }
\description{This is a final project of statistical computation class of SA23204165 which achieves group least angle regression, path plotting and Cp criterion.}
\author{
Jin fulong
Maintainer: Jin fulong <jinfulong@mail.ustc.edu.cn>
}
\keyword{ package }